Fix the way jsConnect handles missing values
Previously, jsConnect would take some of the standard user fields and set them to an empty string if not provided. This would have the side-effect of generating field required errors on subsequent signins which isn’t correct. What should happen is a user should be prompted for the missing field on first sign in and then have subsequent signins work.

To see the effect of this bug:

1. Create a fake jsConnect source with a username, but no email address.
2. SSO with that connection.
3. Change the username in the database or wherever.
4. Sign out and then SSO again.
5. Before the username will not be updated because there will be a hidden error about a missing email address. After this PR the username should update.

As with all SSO PRs this could have regression bugs, but I’m hoping it only affects the small number of SSO scenarios where customers don’t provide one of email or username.